requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
plotly>=5.0.0
watchdog>=4.0.0
